+++ manager/frontend/vite.config.js
@@ -11,12 +11,12 @@ export default defineConfig({
   },
   server: {
     host: '0.0.0.0',
-    // port: 6081,
-    port: 3000,
+    port: 6081,
+    // port: 3000,
     proxy: {
       '/api': {
-        // target: process.env.VITE_API_TARGET || 'http://121.41.200.173:6080',
-        target: process.env.VITE_API_TARGET || 'http://127.0.0.1:8080',
+        target: process.env.VITE_API_TARGET || 'http://121.41.200.173:6080',
+        // target: process.env.VITE_API_TARGET || 'http://127.0.0.1:8080',
         changeOrigin: true
       }
     }